consumer/websocketDashboard/requirements.txt
fastapi
uvicorn[standard]
ssi-fc-data  # nếu bạn dùng thư viện này
kafka-python
confluent-kafka